+++ packages/@headlessui-vue/src/components/combobox/combobox.test.ts
@@ -98,7 +98,7 @@ describe('safeguards', () => {
   ])(
     'should error when we are using a <%s /> without a parent <Combobox />',
     suppressConsoleLogs((name, Component) => {
-      expect(() => render(Component)).toThrowError(
+      expect(() => render(Component)).toThrow(
         `<${name} /> is missing a parent <Combobox /> component.`
       )
     })
@@ -6079,7 +6079,7 @@ describe('Form compatibility', () => {
     // Submit the form
     await click(getByText('Submit'))
 
-    expect(submits).lastCalledWith([['delivery', 'pickup']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'pickup']])
   })
 
   it('should be possible to submit a form with a value', async () => {
@@ -6119,7 +6119,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([]) // no data
+    expect(submits).toHaveBeenLastCalledWith([]) // no data
 
     // Open combobox again
     await click(getComboboxButton())
@@ -6131,7 +6131,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([['delivery', 'home-delivery']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'home-delivery']])
 
     // Open combobox again
     await click(getComboboxButton())
@@ -6143,7 +6143,50 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([['delivery', 'pickup']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'pickup']])
+  })
+
+  it('should not submit the data if the Combobox is disabled', async () => {
+    let submits = jest.fn()
+
+    renderTemplate({
+      template: html`
+        <form @submit="handleSubmit">
+          <input type="hidden" name="foo" value="bar" />
+          <Combobox v-model="value" name="delivery" disabled>
+            <ComboboxInput />
+            <ComboboxButton>Trigger</ComboboxButton>
+            <ComboboxOptions>
+              <ComboboxOption value="pickup">Pickup</ComboboxOption>
+              <ComboboxOption value="home-delivery">Home delivery</ComboboxOption>
+              <ComboboxOption value="dine-in">Dine in</ComboboxOption>
+            </ComboboxOptions>
+          </Combobox>
+          <button>Submit</button>
+        </form>
+      `,
+      setup: () => {
+        let value = ref('home-delivery')
+        return {
+          value,
+          handleSubmit(event: SubmitEvent) {
+            event.preventDefault()
+            submits([...new FormData(event.currentTarget as HTMLFormElement).entries()])
+          },
+        }
+      },
+    })
+
+    // Open combobox
+    await click(getComboboxButton())
+
+    // Submit the form
+    await click(getByText('Submit'))
+
+    // Verify that the form has been submitted
+    expect(submits).toHaveBeenLastCalledWith([
+      ['foo', 'bar'], // The only available field
+    ])
   })
 
   it('should be possible to submit a form with a complex value object', async () => {
@@ -6205,7 +6248,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '1'],
       ['delivery[value]', 'pickup'],
       ['delivery[label]', 'Pickup'],
@@ -6222,7 +6265,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '2'],
       ['delivery[value]', 'home-delivery'],
       ['delivery[label]', 'Home delivery'],
@@ -6239,7 +6282,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '1'],
       ['delivery[value]', 'pickup'],
       ['delivery[label]', 'Pickup'],