+++ packages/@headlessui-vue/src/components/listbox/listbox.test.tsx
@@ -74,7 +74,7 @@ describe('safeguards', () => {
   ])(
     'should error when we are using a <%s /> without a parent <Listbox />',
     suppressConsoleLogs((name, Component) => {
-      expect(() => render(Component)).toThrowError(
+      expect(() => render(Component)).toThrow(
         `<${name} /> is missing a parent <Listbox /> component.`
       )
     })
@@ -5005,7 +5005,7 @@ describe('Form compatibility', () => {
     // Submit the form
     await click(getByText('Submit'))
 
-    expect(submits).lastCalledWith([['delivery', 'pickup']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'pickup']])
   })
 
   it('should be possible to submit a form with a value', async () => {
@@ -5044,7 +5044,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([]) // no data
+    expect(submits).toHaveBeenLastCalledWith([]) // no data
 
     // Open listbox again
     await click(getListboxButton())
@@ -5056,7 +5056,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([['delivery', 'home-delivery']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'home-delivery']])
 
     // Open listbox again
     await click(getListboxButton())
@@ -5068,7 +5068,49 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([['delivery', 'pickup']])
+    expect(submits).toHaveBeenLastCalledWith([['delivery', 'pickup']])
+  })
+
+  it('should not submit the data if the Listbox is disabled', async () => {
+    let submits = jest.fn()
+
+    renderTemplate({
+      template: html`
+        <form @submit="handleSubmit">
+          <input type="hidden" name="foo" value="bar" />
+          <Listbox v-model="value" name="delivery" disabled>
+            <ListboxButton>Trigger</ListboxButton>
+            <ListboxOptions>
+              <ListboxOption value="pickup">Pickup</ListboxOption>
+              <ListboxOption value="home-delivery">Home delivery</ListboxOption>
+              <ListboxOption value="dine-in">Dine in</ListboxOption>
+            </ListboxOptions>
+          </Listbox>
+          <button>Submit</button>
+        </form>
+      `,
+      setup: () => {
+        let value = ref('home-delivery')
+        return {
+          value,
+          handleSubmit(event: SubmitEvent) {
+            event.preventDefault()
+            submits([...new FormData(event.currentTarget as HTMLFormElement).entries()])
+          },
+        }
+      },
+    })
+
+    // Open listbox
+    await click(getListboxButton())
+
+    // Submit the form
+    await click(getByText('Submit'))
+
+    // Verify that the form has been submitted
+    expect(submits).toHaveBeenLastCalledWith([
+      ['foo', 'bar'], // The only available field
+    ])
   })
 
   it('should be possible to submit a form with a complex value object', async () => {
@@ -5129,7 +5171,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '1'],
       ['delivery[value]', 'pickup'],
       ['delivery[label]', 'Pickup'],
@@ -5146,7 +5188,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '2'],
       ['delivery[value]', 'home-delivery'],
       ['delivery[label]', 'Home delivery'],
@@ -5163,7 +5205,7 @@ describe('Form compatibility', () => {
     await click(getByText('Submit'))
 
     // Verify that the form has been submitted
-    expect(submits).lastCalledWith([
+    expect(submits).toHaveBeenLastCalledWith([
       ['delivery[id]', '1'],
       ['delivery[value]', 'pickup'],
       ['delivery[label]', 'Pickup'],