+++ packages/@headlessui-vue/src/components/radio-group/radio-group.ts
@@ -262,6 +262,7 @@ export let RadioGroup = defineComponent({
                   hidden: true,
                   readOnly: true,
                   form,
+                  disabled,
                   name,
                   value,
                 })